Authenticate messages with hmac_secret from DB
A classical mapping was added (db.py), to access the hmac_secret in the auth_user
table of our database.

The argument -s "secret" can now be used to specify the hmac_secret of
the user when starting the client.

The server now extracts the hmac_secret of a connecting user from the auth_user table.

Contributon to issue #1